Подготовка первого теста
