+++ src/Server/Handlers/Adapters/TaskTypeAdapter.java
@@ -1,4 +1,4 @@
-package Handlers.Adapters;
+package Server.Handlers.Adapters;
 
 import com.google.gson.TypeAdapter;
 import com.google.gson.stream.JsonReader;
