+++ src/Server/Handlers/EpicsHandler.java
@@ -1,4 +1,4 @@
-package Handlers;
+package Server.Handlers;
 
 import com.google.gson.Gson;
 import com.google.gson.JsonElement;
@@ -6,20 +6,20 @@ import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
 import com.sun.net.httpserver.HttpExchange;
 import com.sun.net.httpserver.HttpHandler;
+import task.Epic;
 import task.Subtask;
 import taskmanager.TaskManager;
 import java.io.InputStream;
 import java.io.OutputStream;
 import java.nio.charset.StandardCharsets;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
-public class SubtasksHandler extends BaseHttpHandler implements HttpHandler {
-
+public class EpicsHandler extends BaseHttpHandler implements HttpHandler {
 
     TaskManager manager;
-
-    public SubtasksHandler(TaskManager manager) {
+    public EpicsHandler(TaskManager manager) {
         this.manager = manager;
     }
 
@@ -30,11 +30,13 @@ public class SubtasksHandler extends BaseHttpHandler implements HttpHandler {
             Gson gson = createGson();
             String method = exchange.getRequestMethod();
             int pathLength = getPathLength(exchange);
+
+
             switch (method) {
                 case "GET" -> {
                     switch (pathLength) {
                         case 2 -> {
-                            List<Subtask> tasks = manager.getSubtasks();
+                            List<Epic> tasks = manager.getEpicTasks();
                             if (tasks == null) {
                                 sendNotFound(exchange, "Задача не найдена");
                             }
@@ -46,14 +48,26 @@ public class SubtasksHandler extends BaseHttpHandler implements HttpHandler {
                             if (id.isEmpty()) {
                                 sendNotFound(exchange, "Задача не найдена");
                             } else {
-                                Subtask newTask = manager.getSubtask(id.get());
+                                Epic newTask = manager.getEpicTask(id.get());
                                 if (newTask == null) {
                                     sendNotFound(exchange, "Задача не найдена");
                                 }
                             }
-                            Subtask task = manager.getSubtask(id.get());
+                            Epic task = manager.getEpicTask(id.get());
                             String taskJson = gson.toJson(task);
                             sendText(exchange, taskJson);
+                        }
+                        case 4 -> {
+                            Optional<Integer> epicId = getTaskId(exchange);
+                            if (epicId.isEmpty()) {
+                                sendNotFound(exchange, "Задача не найдена");
+                            } else {
+                                if (manager.getEpicTask(epicId.get()) != null) {
+                                    ArrayList<Subtask> subtasks = manager.getEpicSubtasks(epicId.get());
+                                    String subtasksJson = gson.toJson(subtasks);
+                                    sendText(exchange, subtasksJson);
+                                }
+                            }
                         }
                     }
                 }
@@ -65,24 +79,20 @@ public class SubtasksHandler extends BaseHttpHandler implements HttpHandler {
                         sendNotAcceptable(exchange, "Это не Json объект");
                     }
                     JsonObject jsonObject = element.getAsJsonObject();
-                    Subtask taskFromJson = gson.fromJson(jsonObject, Subtask.class);
+                    Epic taskFromJson = gson.fromJson(jsonObject, Epic.class);
                     switch (pathLength) {
                         case 2 -> {
-                            Subtask newTask = manager.addSubtask(taskFromJson);
-                            if (newTask == null) {
-                                sendNotAcceptable(exchange, "Задача пересекается по времени");
-                            } else {
-                                exchange.sendResponseHeaders(201, 0);
-                                exchange.close();
-                            }
+                            manager.addEpicTask(taskFromJson);
+                            exchange.sendResponseHeaders(201, 0);
+                            exchange.close();
                         }
                         case 3 -> {
                             Optional<Integer> id = getTaskId(exchange);
                             if (id.isPresent()) {
-                                Subtask task2 = manager.updateSubtask(taskFromJson);
+                                Epic task2 = manager.updateEpicTask(taskFromJson);
                                 if (task2 != null) {
-                                    exchange.sendResponseHeaders(201, 0);
-                                    exchange.close();
+                                        exchange.sendResponseHeaders(201, 0);
+                                        exchange.close();
                                 } else {
                                     sendNotAcceptable(exchange, "Возникло пересечение");
                                 }
@@ -95,15 +105,16 @@ public class SubtasksHandler extends BaseHttpHandler implements HttpHandler {
                 case "DELETE" -> {
                     switch (pathLength) {
                         case 2 -> {
-                            manager.deleteSubtasks();
+                            manager.deleteEpicTasks();
                             exchange.sendResponseHeaders(201, 0);
                             exchange.close();
+
                         }
                         case 3 -> {
                             Optional<Integer> id = getTaskId(exchange);
                             if (id.isPresent()) {
-                                if (manager.getSubtask(id.get()) != null) {
-                                    manager.deleteSubtask(id.get());
+                                if (manager.getEpicTask(id.get()) != null) {
+                                    manager.deleteEpic(id.get());
                                     exchange.sendResponseHeaders(201, 0);
                                     exchange.close();
                                 } else {
@@ -121,11 +132,11 @@ public class SubtasksHandler extends BaseHttpHandler implements HttpHandler {
                     }
                     exchange.close();
                 }
+
             }
-        } catch (Exception e) {
+        } catch (Exception  e) {
             System.out.println("Возникло исключение");
         }
 
     }
-
 }
