+++ src/Server/Handlers/PrioritizedHandler.java
@@ -1,4 +1,4 @@
-package Handlers;
+package Server.Handlers;
 
 import com.google.gson.Gson;
 import com.sun.net.httpserver.HttpExchange;
