+++ src/Server/Handlers/TasksHandler.java
@@ -1,25 +1,21 @@
-package Handlers;
+package Server.Handlers;
 
-import com.google.gson.Gson;
-import com.google.gson.JsonElement;
-import com.google.gson.JsonObject;
-import com.google.gson.JsonParser;
+import com.google.gson.*;
 import com.sun.net.httpserver.HttpExchange;
 import com.sun.net.httpserver.HttpHandler;
-import task.Epic;
-import task.Subtask;
+import task.Task;
 import taskmanager.TaskManager;
-import java.io.InputStream;
+import java.io.*;
 import java.io.OutputStream;
 import java.nio.charset.StandardCharsets;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
-public class EpicsHandler extends BaseHttpHandler implements HttpHandler {
+public class TasksHandler extends BaseHttpHandler implements HttpHandler {
 
     TaskManager manager;
-    public EpicsHandler(TaskManager manager) {
+
+    public TasksHandler(TaskManager manager) {
         this.manager = manager;
     }
 
@@ -36,7 +32,7 @@ public class EpicsHandler extends BaseHttpHandler implements HttpHandler {
                 case "GET" -> {
                     switch (pathLength) {
                         case 2 -> {
-                            List<Epic> tasks = manager.getEpicTasks();
+                            List<Task> tasks = manager.getSimpleTasks();
                             if (tasks == null) {
                                 sendNotFound(exchange, "Задача не найдена");
                             }
@@ -48,26 +44,14 @@ public class EpicsHandler extends BaseHttpHandler implements HttpHandler {
                             if (id.isEmpty()) {
                                 sendNotFound(exchange, "Задача не найдена");
                             } else {
-                                Epic newTask = manager.getEpicTask(id.get());
+                                Task newTask = manager.getSimpleTask(id.get());
                                 if (newTask == null) {
                                     sendNotFound(exchange, "Задача не найдена");
                                 }
                             }
-                            Epic task = manager.getEpicTask(id.get());
+                            Task task = manager.getSimpleTask(id.get());
                             String taskJson = gson.toJson(task);
                             sendText(exchange, taskJson);
-                        }
-                        case 4 -> {
-                            Optional<Integer> epicId = getTaskId(exchange);
-                            if (epicId.isEmpty()) {
-                                sendNotFound(exchange, "Задача не найдена");
-                            } else {
-                                if (manager.getEpicTask(epicId.get()) != null) {
-                                    ArrayList<Subtask> subtasks = manager.getEpicSubtasks(epicId.get());
-                                    String subtasksJson = gson.toJson(subtasks);
-                                    sendText(exchange, subtasksJson);
-                                }
-                            }
                         }
                     }
                 }
@@ -79,20 +63,24 @@ public class EpicsHandler extends BaseHttpHandler implements HttpHandler {
                         sendNotAcceptable(exchange, "Это не Json объект");
                     }
                     JsonObject jsonObject = element.getAsJsonObject();
-                    Epic taskFromJson = gson.fromJson(jsonObject, Epic.class);
+                    Task taskFromJson = gson.fromJson(jsonObject, Task.class);
                     switch (pathLength) {
                         case 2 -> {
-                            manager.addEpicTask(taskFromJson);
-                            exchange.sendResponseHeaders(201, 0);
-                            exchange.close();
+                            Task newTask = manager.addSimpleTask(taskFromJson);
+                            if (newTask == null) {
+                                sendNotAcceptable(exchange, "Задача пересекается по времени");
+                            } else {
+                                exchange.sendResponseHeaders(201, 0);
+                                exchange.close();
+                            }
                         }
                         case 3 -> {
                             Optional<Integer> id = getTaskId(exchange);
                             if (id.isPresent()) {
-                                Epic task2 = manager.updateEpicTask(taskFromJson);
+                                Task task2 = manager.updateSimpleTask(taskFromJson);
                                 if (task2 != null) {
-                                        exchange.sendResponseHeaders(201, 0);
-                                        exchange.close();
+                                    exchange.sendResponseHeaders(201, 0);
+                                    exchange.close();
                                 } else {
                                     sendNotAcceptable(exchange, "Возникло пересечение");
                                 }
@@ -105,16 +93,15 @@ public class EpicsHandler extends BaseHttpHandler implements HttpHandler {
                 case "DELETE" -> {
                     switch (pathLength) {
                         case 2 -> {
-                            manager.deleteEpicTasks();
+                            manager.deleteSimpleTasks();
                             exchange.sendResponseHeaders(201, 0);
                             exchange.close();
-
                         }
                         case 3 -> {
                             Optional<Integer> id = getTaskId(exchange);
                             if (id.isPresent()) {
-                                if (manager.getEpicTask(id.get()) != null) {
-                                    manager.deleteEpic(id.get());
+                                if (manager.getSimpleTask(id.get()) != null) {
+                                    manager.deleteSimpleTask(id.get());
                                     exchange.sendResponseHeaders(201, 0);
                                     exchange.close();
                                 } else {
@@ -132,9 +119,8 @@ public class EpicsHandler extends BaseHttpHandler implements HttpHandler {
                     }
                     exchange.close();
                 }
-
             }
-        } catch (Exception  e) {
+        } catch (Exception e) {
             System.out.println("Возникло исключение");
         }
